android: Add to launcher button
